+++ packages/google-vertex/src/google-vertex-provider.ts
@@ -1,4 +1,4 @@
-import { LanguageModelV1, ProviderV1 } from '@ai-sdk/provider';
+import { LanguageModelV1, ProviderV1, ImageModelV1 } from '@ai-sdk/provider';
 import {
   FetchFunction,
   generateId,
@@ -16,6 +16,10 @@ import {
 } from './google-vertex-embedding-settings';
 import { GoogleVertexEmbeddingModel } from './google-vertex-embedding-model';
 import { GoogleGenerativeAILanguageModel } from '@ai-sdk/google/internal';
+import {
+  GoogleVertexImageModel,
+  GoogleVertexImageModelId,
+} from './google-vertex-image-model';
 
 export interface GoogleVertexProvider extends ProviderV1 {
   /**
@@ -30,6 +34,11 @@ Creates a model for text generation.
     modelId: GoogleVertexModelId,
     settings?: GoogleVertexSettings,
   ) => LanguageModelV1;
+
+  /**
+   * Creates a model for image generation.
+   */
+  image(modelId: GoogleVertexImageModelId): ImageModelV1;
 }
 
 export interface GoogleVertexProviderSettings {
@@ -123,6 +132,14 @@ export function createVertex(
       baseURL: loadBaseURL(),
     });
 
+  const createImageModel = (modelId: GoogleVertexImageModelId) =>
+    new GoogleVertexImageModel(modelId, {
+      provider: `google.vertex.image`,
+      baseURL: loadBaseURL(),
+      headers: options.headers ?? {},
+      fetch: options.fetch,
+    });
+
   const provider = function (
     modelId: GoogleVertexModelId,
     settings?: GoogleVertexSettings,
@@ -138,6 +155,7 @@ export function createVertex(
 
   provider.languageModel = createChatModel;
   provider.textEmbeddingModel = createEmbeddingModel;
+  provider.image = createImageModel;
 
   return provider as GoogleVertexProvider;
 }