+++ packages/google/src/google-generative-ai-settings.ts
@@ -55,6 +55,12 @@ Optional. A list of unique safety settings for blocking unsafe content.
       | 'BLOCK_ONLY_HIGH'
       | 'BLOCK_NONE';
   }>;
+  /**
+   * Optional. Enables timestamp understanding for audio-only files.
+   *
+   * https://cloud.google.com/vertex-ai/generative-ai/docs/multimodal/audio-understanding
+   */
+  audioTimestamp?: boolean;
 
   /**
 Optional. When enabled, the model will use Google search to ground the response.