+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,21 @@
 
 ### Patch Changes
 
+- Updated dependencies [da9d240]
+- Updated dependencies [6f1bfde]
+  - ai@4.0.20
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [c3a6065]
+  - ai@4.0.19
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [304e6d3]
 - Updated dependencies [304e6d3]
   - ai@4.0.18