+++ packages/openai-compatible/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/openai-compatible
 
+## 0.0.9
+
+### Patch Changes
+
+- 9c7653b: feat (docs): Update OpenAI Compatible docs for new package.
+
+## 0.0.8
+
+### Patch Changes
+
+- 6faab13: feat (provider/openai-compatible): simulated streaming setting
+
+## 0.0.7
+
+### Patch Changes
+
+- ad2bf11: feat (provider/fireworks): Add Fireworks provider.
+
 ## 0.0.6
 
 ### Patch Changes